refactor(front): Pass A assistant gotchas — move DTOs + pure local types to types/api
All 3 files' local domain types were pure (no @app/lib coupling), so each moves
the DTO plus its local type(s) to types/api.

- assistant/plan_mode: PlanApprovalState + GetConversationPlanModeResponseBody.
- assistant/observability/overview: AgentCostStats + GetAgentOverviewResponseBody
  (lib back-imports AgentCostStats).
- assistant/conversation/attachments: full attachment DTO tree
  (ConversationAttachmentType and friends); lib back-imports the 6 used by kept
  serializer/guard functions. Also fixes the ConversationAttachmentType import in
  the deferred projects/context.ts.
- 31 consumers updated (split-aware). No deferrals.
