+++ front/lib/api/actions/servers/pod_manager/build_pod_search_data_sources.ts
@@ -1,7 +1,6 @@
 import { getDataSourceURI } from "@app/lib/actions/mcp_internal_actions/input_configuration";
 import type { DataSourcesToolConfigurationType } from "@app/lib/actions/mcp_internal_actions/input_schemas";
 import type { DataSourceConfiguration } from "@app/lib/api/assistant/configuration/types";
-import type { ContentNodeAttachmentType } from "@app/lib/api/assistant/conversation/attachments";
 import {
   isContentFragmentDataSourceNode,
   isContentNodeAttachmentType,
@@ -16,6 +15,7 @@ import {
 } from "@app/lib/api/projects/data_sources";
 import type { Authenticator } from "@app/lib/auth";
 import type { SpaceResource } from "@app/lib/resources/space_resource";
+import type { ContentNodeAttachmentType } from "@app/types/api/assistant/conversation/attachments";
 import { INTERNAL_MIME_TYPES } from "@dust-tt/client";
 
 export type PodSemanticSearchScope = "files" | "conversations" | "all";
