+++ front/lib/api/assistant/plan_mode.ts
@@ -2,15 +2,7 @@ import { PLAN_MODE_SKELETON } from "@app/lib/api/actions/servers/plan_mode/metad
 import type { Authenticator } from "@app/lib/auth";
 import { executeWithLock } from "@app/lib/lock";
 import { FileResource } from "@app/lib/resources/file_resource";
-import type { FileType, PlanModeApproval } from "@app/types/files";
-
-export type PlanApprovalState = "draft" | "pending" | "approved";
-
-export type GetConversationPlanModeResponseBody = {
-  planFile: FileType | null;
-  content: string | null;
-  approvalState: PlanApprovalState;
-};
+import type { PlanModeApproval } from "@app/types/files";
 
 // Conversation-scoped lock for all plan-mode operations. One plan per conversation, so a single
 // lock keyed on conversation sId serializes create/edit/approve/close and prevents races (e.g.
